+++ optim_esm_tools/region_finding/__init__.py
@@ -1,9 +1,9 @@
-from ._base import MaskAll
+from .keep_all import MaskAll
 from .iter_ranges import *
 from .local_history import LocalHistory
 from .max_region import MaxRegion
 from .named_region import Asia
-from .named_region import NamedRegions
+from .named_region import Medeteranian
 from .percentiles import Percentiles
 from .percentiles_history import PercentilesHistory
 from .product_percentiles import ProductPercentiles
